rbd: squash ImportDiff.cc into Import.cc.
